# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MAIS Wellbest API Documentation",
        "description": "Standard OpenAPI documentation for the MAIS Wellbest application.",
        "contact": {
            "email": "admin@rexonindo.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "MAIS Wellbest API Server"
        }
    ],
    "paths": {
...
        "/work-orders": {
            "get": {
                "tags": [
                    "Transactions - Work Orders"
                ],
                "summary": "Get list of work orders",
                "description": "Returns a paginated list of work orders. Can be filtered by status.",
                "operationId": "getWorkOrdersList",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by active status ('Planned', 'In Progress').",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/WorkOrder"
                                            }
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Transactions - Work Orders"
                ],
                "summary": "Create a new work order",
                "description": "Stores a new work order and returns the created work order data.",
                "operationId": "storeWorkOrder",
                "requestBody": {
                    "description": "Work order data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreWorkOrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Work order created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WorkOrder"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/work-orders/{id}": {
            "get": {
                "tags": [
                    "Transactions - Work Orders"
                ],
                "summary": "Get work order information",
                "description": "Returns work order data by ID.",
                "operationId": "getWorkOrderById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Work Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WorkOrder"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Transactions - Work Orders"
                ],
                "summary": "Update existing work order",
                "description": "Updates an existing work order and returns the updated work order data.",
                "operationId": "updateWorkOrder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Work Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Work order data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreWorkOrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Work order updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WorkOrder"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Transactions - Work Orders"
                ],
                "summary": "Delete existing work order",
                "description": "Deletes a work order record and returns no content.",
                "operationId": "deleteWorkOrder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Work Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/work-orders/{work_order}/valid-processes": {
            "get": {
                "tags": [
                    "Transactions - Work Orders"
                ],
                "summary": "Get valid processes for a Work Order",
                "description": "Helper endpoint to get a list of valid processes for a given Work Order, based on its item type.",
                "operationId": "getValidProcessesForWo",
                "parameters": [
                    {
                        "name": "work_order",
                        "in": "path",
                        "description": "The `wo_no` of the work order.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of valid processes.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Process"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Work Order not found."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "PaginationLinks": {
                "title": "Pagination Links",
                "properties": {
                    "first": {
                        "type": "string",
                        "example": "http://localhost/api/v1/departments?page=1"
                    },
                    "last": {
                        "type": "string",
                        "example": "http://localhost/api/v1/departments?page=5"
                    },
                    "prev": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "example": "http://localhost/api/v1/departments?page=2",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "title": "Pagination Meta",
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "from": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 5
                    },
                    "path": {
                        "type": "string",
                        "example": "http://localhost/api/v1/departments"
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "to": {
                        "type": "integer",
                        "example": 15
                    },
                    "total": {
                        "type": "integer",
                        "example": 75
                    }
                },
                "type": "object"
            },
            "Department": {
                "title": "Department",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "departmentCode": {
                        "type": "string",
                        "example": "PROD"
                    },
                    "name": {
                        "type": "string",
                        "example": "Production"
                    },
                    "description": {
                        "type": "string",
                        "example": "Main production department",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Item": {
                "title": "Item",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "itemCode": {
                        "type": "string",
                        "example": "FG-001"
                    },
                    "name": {
                        "type": "string",
                        "example": "Finished Good 1"
                    },
                    "description": {
                        "type": "string",
                        "example": "Main finished good",
                        "nullable": true
                    },
                    "itemType": {
                        "type": "string",
                        "example": "FINISH GOOD"
                    },
                    "uom": {
                        "type": "string",
                        "example": "PCS"
                    }
                },
                "type": "object"
            },
...
            "Process": {
                "title": "Process",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "processCode": {
                        "type": "string",
                        "example": "PROC-01"
                    },
                    "name": {
                        "type": "string",
                        "example": "Cutting"
                    },
                    "description": {
                        "type": "string",
                        "example": "Process of cutting raw materials",
                        "nullable": true
                    },
                    "department": {
                        "$ref": "#/components/schemas/Department"
                    }
                },
                "type": "object"
            },
...
            "WorkOrder": {
                "title": "Work Order",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "workOrderNumber": {
                        "type": "string",
                        "example": "WO-2025-001"
                    },
                    "orderDate": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-10-26"
                    },
                    "planDate": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-10-27"
                    },
                    "planQty": {
                        "type": "number",
                        "format": "float",
                        "example": 100.5
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "Planned",
                            "In Progress",
                            "Completed",
                            "Cancelled"
                        ],
                        "example": "Planned"
                    },
                    "item": {
                        "$ref": "#/components/schemas/Item"
                    }
                },
                "type": "object"
            },
            "StoreWorkOrderRequest": {
                "title": "Store Work Order Request",
                "required": [
                    "wo_no",
                    "itm_cd",
                    "ord_dt",
                    "plan_dt",
                    "plan_qty"
                ],
                "properties": {
                    "wo_no": {
                        "type": "string",
                        "maxLength": 20,
                        "example": "WO-2025-002"
                    },
                    "itm_cd": {
                        "type": "string",
                        "maxLength": 20,
                        "example": "FG-001"
                    },
                    "ord_dt": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-10-28"
                    },
                    "plan_dt": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-10-29"
                    },
                    "plan_qty": {
                        "type": "number",
                        "format": "float",
                        "example": 250
                    },
                    "stats": {
                        "type": "string",
                        "enum": [
                            "Planned",
                            "In Progress",
                            "Completed",
                            "Cancelled"
                        ],
                        "example": "Planned"
                    },
                    "rmks": {
                        "type": "string",
                        "example": "Urgent order",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}